+++ Funcky/Extensions/Including.cs
@@ -1,13 +1,19 @@
+#if GENERIC_MATH
+using System.Numerics;
+
 namespace Funcky.Extensions;
 
-public class Including : IIntervalBoundary
+/// <summary>An interval boundary which includes its value.</summary>
+public readonly struct Including : IIntervalBoundary
 {
-    private Including(int number)
-    {
-        Value = number;
-    }
+    [Pure]
+    public static bool IsLowerBoundOf<T>(T boundary, T number)
+        where T : IComparisonOperators<T, T, bool>
+        => boundary <= number;
 
-    public int Value { get; }
-
-    public static implicit operator Including(int number) => new(number);
+    [Pure]
+    public static bool IsUpperBoundOf<T>(T boundary, T number)
+        where T : IComparisonOperators<T, T, bool>
+        => number <= boundary;
 }
+#endif